OpenJDK7 support(first step)
So this patch adds the openjdk7 parameter for --with-java-runtime-library configuration calls.
Next to this, hpi is removed because openjdk7 doesn't suppport it anymore.
Functionallity got implemented in cacao in most cases.
Class java.lang.Thread changed slightly.
You'll have to fix some things in icedtea to build it with version 7.
the c++ class for java_lang_reflect_Field and java_lang_relfect_Constructor got a new member, as the AssesibleObject it extends got a new member
Startup of the cacaovm and "Hello World" for x86_64 supported.

* configure.ac : hpi has to be removed
* m4/java-runtime-library.m4 : adding openjdk7 as an option for
														./configure --with-java-runtime-library
* m4/jmm.m4 : include openjdk7 parameter
* m4/jvm.m4 : likewise
* src/native/vm/nativevm.cpp : Kick out hpi for openjdk7
* src/native/vm/openjdk/hpi.cpp : likewise
* src/native/vm/openjdk/jvm.cpp : likewise
																	All methods formerly using hpi to implement their functionality are now implemented by cacao.
* src/threads/thread-openjdk.cpp : java_lang_Thread.me is removed for openjdk7
* src/vm/global.h : openjdk7 is more strict about versions
* src/vm/javaobjects.cpp : java_lang_Thread.me is removed for openjdk7
* src/vm/javaobjects.hpp : likewise and
													 java_lang_reflect_Field has new member securityCheckCache of type Object
													 java_lang_reflect_Constructor also moved its securityCheckCache member
													 java_lang_reflect_Method changed likewise
* src/vm/vm.hpp : Kick out hpi for openjdk7