Made some exception messages more hotspot like.
The message generated for a UnsatisfiedLinkError, ClassCastException and OutOfMemoryError with OpenJDK
are now more alike to those generated by hotspot.

ClassCastException still does not tell you with class you tried to cast to, only what you tried to cast from.
This is because the trap logic does not give us enough information as of yet.