compiler2: Provide x86_64 lowering logic for INVOKESTATICInst
Lowering is now possible for INVOKESTATICInst with int/long/float/double
operands. However, there are a few things that still have to be addressed due
to the pending refactoring of the patching mechanisms for the compiler2.
First of all, lowering of invocations to unresolved methods is not yet
supported. Additionally, for invocations to resolved methods, we optimistically
assume that the entrypoint to the callee is already known - this is unsafe in
general, since we actually would have to place a patcher that ensures that the
callee has already been compiled and patches the callee's entrypoint address at
the call-site.

* src/vm/jit/compiler2/x86_64/X86_64Backend.cpp: Add lowering logic.
* tests/compiler2/junit/org/cacaojvm/compiler2/test/All.java: Add tests.
* tests/compiler2/junit/org/cacaojvm/compiler2/test/InvokeStatic.java: Add tests.