Add option `-XX:ReplaceMethod` to restrict OSR to a single method
The option can be used from the command line as `-XX:ReplaceMethod=<value>` and
assures that on-stack replacement is only performed for the specified method.
For example, with `-XX:ReplaceMethod=foo` on-stack replacement would only be
triggered for the method named `foo` (i.e. the matching is only performed on
the name of the method, ignoring its signature, etc.).

In general, the on-stack replacement is driven by "countdown traps" - special
instrumentation code that is placed in methods compiled by the baseline compiler
to estimate the execution frequency of methods, based on a countdown. As soon
as a method's countdown reaches a certain threshold, the on-stack replacement
mechanism will be triggered to recompile the method with the compiler2 and
execution will be transferred to the optimized code (NOTE: this transition is
not yet fully implemented). This means in order to prevent a method from being
recompiled and replaced via OSR it is sufficient not to insrument the method.
This is exactly what we do when using the new option, i.e. with
`-XX:ReplaceMethod=foo` only the method `foo` will be instrumented with
countdown traps, all other methods will be compiled without any insrumentation.

* src/vm/jit/jit.hpp: Add some documentation.
* src/vm/jit/jit.cpp: Mark methods that should be instrumented.
* src/vm/options.hpp: Add option `-XX:ReplaceMethod`.
* src/vm/options.cpp: Likewise.