compiler2: Generate method header
Store a pointer to the method's `codeinfo` structure at the end of its data
segment - this section of the data segment, which solely consists of this
`codeinfo` pointer, is referred to as "method header".

Furthermore, the linking step of the code memory requires that the data segment
is in its final shape and is relocated accordingly.

* src/vm/jit/compiler2/CodeGenPass.cpp: See description above.