compiler2: Temporarily remove AssumptionAnchoringPass
The responsibility of the AssumptionAnchoringPass is to ensure that an
AssumptionInst does not leave the branch containing the program region that
has to be guarded against illegal execution. This is necessary to avoid that
deoptimization is triggered prematurely on a branch that does not even depend on
the corresponding assumption. However, This also requires adaptions to the
scheduling algorithms (e.g. ScheduleLate), therefore for now we remove the
AssumptionAnchoringPass, and leave it for future investigation.