compiler2: Let Address provide type information about referenced data
This enables Address (and its subclasses MachineAddress, NativeAddres, and
X86_64ModRMOperand) to provide the type of data that they reference. So far,
Addresses have always been initialized to be of type Type::ReturnAddressType,
which did not really serve as a useful abstraction.

Furhtermore, the construction of MoveInsts at various places in
X86_64Backend.cpp is refactored to reduce code duplication.